+++ src/forwarding.rs
@@ -135,7 +135,7 @@ impl ForwardingError {
 /// as the value.
 pub async fn serve(
     mut wormhole: Wormhole,
-    transit_handler: impl FnOnce(transit::TransitInfo, std::net::SocketAddr),
+    transit_handler: impl FnOnce(transit::TransitInfo),
     relay_hints: Vec<transit::RelayHint>,
     targets: Vec<(Option<url::Host>, u16)>,
     cancel: impl Future<Output = ()>,
@@ -190,7 +190,7 @@ pub async fn serve(
         },
     };
 
-    let (mut transit, info, addr) = match connector
+    let (mut transit, info) = match connector
         .leader_connect(
             wormhole.key().derive_transit_key(wormhole.appid()),
             peer_version.transit_abilities,
@@ -207,7 +207,7 @@ pub async fn serve(
             return Err(error);
         },
     };
-    transit_handler(info, addr);
+    transit_handler(info);
 
     /* We got a transit, now close the Wormhole */
     wormhole.close().await?;
@@ -518,7 +518,7 @@ impl ForwardingServe {
 /// no more than 1024 ports may be forwarded at once.
 pub async fn connect(
     mut wormhole: Wormhole,
-    transit_handler: impl FnOnce(transit::TransitInfo, std::net::SocketAddr),
+    transit_handler: impl FnOnce(transit::TransitInfo),
     relay_hints: Vec<transit::RelayHint>,
     bind_address: Option<std::net::IpAddr>,
     custom_ports: &[u16],
@@ -561,7 +561,7 @@ pub async fn connect(
         },
     };
 
-    let (mut transit, info, addr) = match connector
+    let (mut transit, info) = match connector
         .follower_connect(
             wormhole.key().derive_transit_key(wormhole.appid()),
             peer_version.transit_abilities,
@@ -578,7 +578,7 @@ pub async fn connect(
             return Err(error);
         },
     };
-    transit_handler(info, addr);
+    transit_handler(info);
 
     /* We got a transit, now close the Wormhole */
     wormhole.close().await?;