+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Added compilation support for WASM targets.
+- \[lib\]\[breaking\] replaced `transit::TransitInfo` with a struct containing the address, the old enum has been renamed to `transit::ConnectionType`.
+
 ## Version 0.6.0
 
 - Add shell completion support for the CLI